+++ src/test/java/application/FakeBurpCertificateBuilder.java
@@ -7,6 +7,7 @@ import java.security.NoSuchProviderException;
 import java.security.spec.InvalidKeySpecException;
 import java.security.spec.RSAPrivateCrtKeySpec;
 import java.security.spec.RSAPublicKeySpec;
+import model.BurpCertificateBuilder;
 
 class FakeBurpCertificateBuilder extends BurpCertificateBuilder {
